backend/requirements.txt
django
djangorestframework
psycopg2
django-cors-headers
python-dotenv
Pillow
djangorestframework-simplejwt